==> code/02b_state_trifectas.Rmd <==
title: "Party Control of State Government"
author: "Daniel Posthumus"
date: 12/28/2024
output: pdf_document
---
Taken from Andy Greens github

Andy Green

11/12/19

This file contains the code I used to build a dataset aimed at analyzing the relationship between party control of state government and a variety of well-being metrics at the state level. As the dataset encompasses data from a variety of different sources, this document is split up such that each of the primary metrics is given a sub-section. The source of the data, including any relevant instructions on how to query the exact data I used, are included under each sub-section heading.


## Set-up:

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)

## Install packages
#install.packages('tidyverse')
#install.packages('rvest')

## Load packages
library(tidyverse)
library(rvest)
library(dplyr)
library(httr)

## Clear objects from prior session
rm(list = ls(all = TRUE))
## setwd("~/Honors_Thesis/SURA/dataset_inputs")
```


## Compiling the state-level party control dataset


Source: I scraped the data from each state's invidividual "Party Control of State Government" page on Ballotpedia. The URL for each state, as seen in the loop below, is given by: https://ballotpedia.org/Party_control_of_<insert state name here>_state_government.


```{r}
#### Looping through each of the states on Ballotpedia ####
## Create inital blank dataset to store everything in 
df <- data.frame()

## Setting up the list of state names for looping. Nebraska has to be removed because they have a unicameral legislature, and thus their table is formatted differently on the web page. I'm collecting the data for them separately outside of the loop. The other state names are being altered to add an underscore such that the URL's will function correctly. These will be removed later.
state.name2 <- as.list(state.name)
state.name2[27] <- NULL
state.name2[28] <- "New_Hampshire"
state.name2[29] <- "New_Jersey"
state.name2[30] <- "New_Mexico"
state.name2[31] <- "New_York"
state.name2[32] <- "North_Carolina"
state.name2[33] <- "North_Dakota"
state.name2[38] <- "Rhode_Island"
state.name2[39] <- "South_Carolina"
state.name2[40] <- "South_Dakota"
state.name2[47] <- "West_Virginia"

## Creating a list to append all the individual state dataframes to
datalist = list()

## Looping through each state
for (state in state.name2) {
  url <- paste("https://ballotpedia.org/Party_control_of_", state, "_state_government", sep = "")
  cat("Processing URL:", url, "\n")
  
  response <- tryCatch({
    GET(url, user_agent("Mozilla/5.0"))
  }, error = function(e) {
    cat("GET request failed for URL:", url, "\nError:", e$message, "\n")
    return(NULL)
  })
  
  if (is.null(response)) next
  
  webpage <- tryCatch({
    read_html(response)
  }, error = function(e) {
    cat("Failed to parse URL:", url, "\nError:", e$message, "\n")
    return(NULL)
  })
  
  if (is.null(webpage)) next
    ## Extracting the table
    data <- html_nodes(webpage, "table.wikitable")
    data <- html_table(data)
    
    ## Converting to dataframe
    data <- as.data.frame(data)
    
    ## Transposing rows and columns
    data <- t(data)
    
    ## Fixing column names (currently stored in first row)
    colnames(data) <- as.character(unlist(data[1,]))
    data <- data[-1, ]
    
    ## Pulling out the year from the rownames
    data2 <- as.data.frame(rownames(data))
    names(data2) <- "year_end"
    data <- cbind(data, data2)
    data$year_end <- sub('.','', data$year_end)
    
    ## Fixing the column name for the House, as some states call it "Assembly"
    names(data)[3]<-"House"
    
    ## Filling in the state name for all rows
    data$state <- state
    
    ## Converting the year values into the full year value
    data$year_beg[data$year_end > 50] <- 19
    data$year_beg[data$year_end < 50] <- 20
    data$year <- paste(data$year_beg, data$year_end, sep = "")
    
    ## Getting rid of the underscores in the state names
    data$state <- sub('_',' ', data$state)
        
    ## Creating a variable that concatenates state and year
    data$stateyear <- paste(data$state,data$year, sep = "")
    
    ## take care of potential [] coming along 
    ## just checks if the variable is of length one and then fills it 
    data <- data %>% 
      mutate( 
        Governor = if_else(nchar(Governor) > 1, strtrim(Governor, 1), Governor),
        House = if_else(nchar(House) > 1, strtrim(House, 1), House),
        Senate = if_else(nchar(Senate) > 1, strtrim(Senate, 1), Senate)
      )
        
    ## Summarizing the total government control
    data$total_gov[data$Governor == "D" & data$Senate == "D" & data$House == "D"] <- "D"
    data$total_gov[data$Governor == "R" & data$Senate == "R" & data$House == "R"] <- "R"
    data$total_gov[data$Governor == "D" & data$Senate == "D" & data$House == "R"] <- "Split"
    data$total_gov[data$Governor == "D" & data$Senate == "R" & data$House == "D"] <- "Split"
    data$total_gov[data$Governor == "D" & data$Senate == "R" & data$House == "R"] <- "Split"
    data$total_gov[data$Governor == "R" & data$Senate == "D" & data$House == "R"] <- "Split"
    data$total_gov[data$Governor == "R" & data$Senate == "R" & data$House == "D"] <- "Split"
    data$total_gov[data$Governor == "R" & data$Senate == "D" & data$House == "D"] <- "Split"
    data$total_gov[data$Governor == "I" & data$Senate == "S" & data$House == "D"] <- "Split"
    data$total_gov[data$Governor == "I" & data$Senate == "R" & data$House == "D"] <- "Split"
    data$total_gov[data$Governor == "I" & data$Senate == "R" & data$House == "R"] <- "Split"
    data$total_gov[data$Governor == "R" & data$Senate == "R" & data$House == "S"] <- "Split"
    data$total_gov[data$Governor == "R" & data$Senate == "S" & data$House == "R"] <- "Split"
    data$total_gov[data$Governor == "R" & data$Senate == "D" & data$House == "S"] <- "Split"
    data$total_gov[data$Governor == "I" & data$Senate == "D" & data$House == "D"] <- "Split"
    data$total_gov[data$Governor == "D" & data$Senate == "S" & data$House == "D"] <- "Split"
    data$total_gov[data$Governor == "D" & data$Senate == "S" & data$House == "R"] <- "Split"
    data$total_gov[data$Governor == "I" & data$Senate == "D" & data$House == "R"] <- "Split"
    data$total_gov[data$Governor == "D" & data$Senate == "D" & data$House == "S"] <- "Split"
    data$total_gov[data$Governor == "D" & data$Senate == "R" & data$House == "S"] <- "Split"
    data$total_gov[data$Governor == "R" & data$Senate == "S" & data$House == "D"] <- "Split"
    ## to deal with split chambers 
    data$total_gov[data$Senate == "S" | data$House == "S"] <- "Split"

    ## add existing data back to big dataframe
    df <- rbind(df, data)
}
unique(df$state)
```

#### Creating the Nebraska dataset ####
```{r}
## Setting the URL for Nebraska
url <- "https://ballotpedia.org/Party_control_of_Nebraska_state_government"

## Reading in the webpage
webpage <- read_html(url)

## Extracting the table
data <- html_nodes(webpage, "table.wikitable")
data <- html_table(data)

## Converting to dataframe
data <- as.data.frame(data)

## Transposing rows and columns
data <- t(data)

## Fixing column names (currently stored in first row)
colnames(data) <- as.character(unlist(data[1,]))
data <- data[-1, ]

## Converting to dataframe
data <- as.data.frame(data)

## Adding in House column (needs to match up with other states)
data$House <- "-"

## Pulling out the year from the rownames
data2 <- as.data.frame(rownames(data))
names(data2) <- "year_end"
data <- cbind(data, data2)
data$year_end <- sub('.','', data$year_end)

## Filling in the state name for all rows
data$state <- "Nebraska"

## Converting the year values into the full year value
data$year_beg[data$year_end > 50] <- 19
data$year_beg[data$year_end < 50] <- 20
data$year <- paste(data$year_beg, data$year_end, sep = "")
    
## Creating a variable that concatenates state and year
data$stateyear <- paste(data$state,data$year, sep = "")

## Adding the columns for total government control and detailed government control
data$total_gov <- "-"

df <- rbind(df, data)

####

```

## Clean up df for export 
```{r} 
df2 <- df %>% 
  select("state", "year", "Governor", "Senate", "House", "total_gov") %>% 
  rename(Year = year, State = state) %>% 
  mutate(
    Year = as.numeric(Year)
  ) %>% 
  filter(Year >= 2010)


``` 

```{r} 
setwd("~/medicaid_project/data/clean/state_level")
write.csv(df2, "state_trifectas_ballotpedia_scrape.csv")

```
